--- src/ESPCN.py
@@ -1,78 +1,0 @@
-#based on code from https://keras.io/examples/vision/super_resolution_sub_pixel/
-import tensorflow as tf
-
-import os
-import math
-import numpy as np
-
-from tensorflow import keras
-from tensorflow.keras import layers
-from tensorflow.keras.preprocessing.image import load_img
-from tensorflow.keras.preprocessing.image import array_to_img
-from tensorflow.keras.preprocessing.image import img_to_array
-from tensorflow.keras.preprocessing import image_dataset_from_directory
-
-def get_model(upscale_factor=3, channels=1):
-    conv_args = {
-        "activation": "relu",
-        "kernel_initializer": "Orthogonal",
-        "padding": "same",
-    }
-    inputs = keras.Input(shape=(None, None, channels))
-    x = layers.Conv2D(64, 5, **conv_args)(inputs)
-    x = layers.Conv2D(64, 3, **conv_args)(x)
-    x = layers.Conv2D(32, 3, **conv_args)(x)
-    x = layers.Conv2D(channels * (upscale_factor ** 2), 3, **conv_args)(x)
-    outputs = tf.nn.depth_to_space(x, upscale_factor)
-
-    return keras.Model(inputs, outputs)
-
-class ESPCNCallback(keras.callbacks.Callback):
-    def __init__(self):
-        super(ESPCNCallback, self).__init__()
-        self.test_img = get_lowres_image(load_img(test_img_paths[0]), upscale_factor)
-
-    # Store PSNR value in each epoch.
-    def on_epoch_begin(self, epoch, logs=None):
-        self.psnr = []
-
-    def on_epoch_end(self, epoch, logs=None):
-        print("Mean PSNR for epoch: %.2f" % (np.mean(self.psnr)))
-        if epoch % 20 == 0:
-            prediction = upscale_image(self.model, self.test_img)
-            plot_results(prediction, "epoch-" + str(epoch), "prediction")
-
-    def on_test_batch_end(self, batch, logs=None):
-        self.psnr.append(10 * math.log10(1 / logs["loss"]))
-
-early_stopping_callback = keras.callbacks.EarlyStopping(monitor="loss", patience=10)
-
-checkpoint_filepath = "/tmp/checkpoint"
-
-model_checkpoint_callback = keras.callbacks.ModelCheckpoint(
-    filepath=checkpoint_filepath,
-    save_weights_only=True,
-    monitor="loss",
-    mode="min",
-    save_best_only=True,
-)
-
-model = get_model(upscale_factor=upscale_factor, channels=1)
-model.summary()
-
-callbacks = [ESPCNCallback(), early_stopping_callback, model_checkpoint_callback]
-loss_fn = keras.losses.MeanSquaredError()
-optimizer = keras.optimizers.Adam(learning_rate=0.001)
-
-epochs = 100
-
-model.compile(
-    optimizer=optimizer, loss=loss_fn,
-)
-
-model.fit(
-    train_ds, epochs=epochs, callbacks=callbacks, validation_data=valid_ds, verbose=2
-)
-
-# The model weights (that are considered the best) are loaded into the model.
-model.load_weights(checkpoint_filepath)
